requirements.txt
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4
scikit-learn==1.5.1
pytorch==2.4.0
pytorch-cuda==12.1
torchvision==0.19.0
tqdm==4.66.4
seaborn==0.13.2